@@ -16,14 +16,15 @@
 
 package com.example.android.persistence.codelab.db;
 
-import android.arch.persistence.room.Dao;
-import android.arch.persistence.room.Delete;
-import android.arch.persistence.room.Insert;
-import android.arch.persistence.room.Query;
 
+import static androidx.room.OnConflictStrategy.IGNORE;
+
+import androidx.room.Dao;
+import androidx.room.Delete;
+import androidx.room.Insert;
+import androidx.room.Query;
 import java.util.List;
 
-import static android.arch.persistence.room.OnConflictStrategy.IGNORE;
 
 @Dao
 public interface UserDao {
